+++ src/util/ArraySet.java
@@ -158,7 +158,8 @@ public final class ArraySet<T> implements Set<T> {
 	public boolean retainAll(@Nullable final Collection<?> coll) {
 		final boolean retval = impl.retainAll(coll);
 		if (retval) {
-			hash = StreamSupport.stream(coll.spliterator(), false).collect(Collectors.summingInt(Object::hashCode));
+			hash = StreamSupport.stream(coll.spliterator(), false).collect(Collectors.summingInt(Object::hashCode))
+					       .intValue();
 		}
 		return retval;
 	}
@@ -171,7 +172,8 @@ public final class ArraySet<T> implements Set<T> {
 	public boolean removeAll(@Nullable final Collection<?> coll) {
 		final boolean retval = impl.removeAll(coll);
 		if (retval) {
-			hash = StreamSupport.stream(coll.spliterator(), false).collect(Collectors.summingInt(Object::hashCode));
+			hash = StreamSupport.stream(coll.spliterator(), false).collect(Collectors.summingInt(Object::hashCode))
+					       .intValue();
 		}
 		return retval;
 	}
