+++ src/view/map/main/ComponentMouseListener.java
@@ -108,7 +108,7 @@ public final class ComponentMouseListener extends MouseAdapter implements
 	 * @return them concatenated together
 	 */
 	private static String concat(final String... strings) {
-		final int len = Stream.of(strings).collect(Collectors.summingInt(String::length));
+		final int len = Stream.of(strings).collect(Collectors.summingInt(String::length)).intValue();
 		final StringBuilder build = new StringBuilder(len);
 		Stream.of(strings).forEach(build::append);
 		return NullCleaner.assertNotNull(build.toString());
